Split JsonFileSolarSystemRepo into two files.
