+++ kuiper-game/Repositories/ISolarSystemJsonStructureRepository.cs
@@ -1,0 +1,7 @@
+namespace Kuiper.Repositories
+{
+    public interface ISolarSystemJsonStructureRepository
+    {
+         string GetSolarSystemJsonStructure();
+    }
+}
